refactor: divided admin and user routes

+++ src/controllers/user.js
@@ -32,9 +32,14 @@ export const getUserById = async (id) => {
 /**
  *
  * @param {string} id
+ * @param {object} user
+ * @param {'admin' | 'seller' | 'customer'} user.rol
  * @returns {Promise<boolean>}
  */
-export const removeUserById = async (id) => {
+export const removeUserById = async (id, user) => {
+  if (!user || user.rol !== 'admin') {
+    throw new Error('You dont have permission')
+  }
   await User.deleteOne({ _id: id })
 
   return true
